requirements.txt
# Required
pypeg2>=2.15.2
django-translated-fields==0.11.3
model-bakery==1.3.3

# Don't include these two below as they are already included in tox.ini
# Django>=1.11
# djangorestframework>=3.5

# Optional
django-filter

# Code style
flake8>=3.7.9
flake8-tidy-imports>=4.1.0
pycodestyle>=2.5.0